Changed Integer.MAX_VALUE to Float.MAX_VALUE

+++ gemp-swccg-logic/src/main/java/com/gempukku/swccgo/logic/modifiers/ModifiersLogic.java
@@ -9382,7 +9382,7 @@ public class ModifiersLogic implements ModifiersEnvironment, ModifiersQuerying, 
         }
 
         // See if we are capped at a value and apply the cap
-        float immunityValueCap = Integer.MAX_VALUE;
+        float immunityValueCap = Float.MAX_VALUE;
         for (Modifier modifier : getModifiersAffectingCard(gameState, ModifierType.IMMUNITY_TO_ATTRITION_LIMITED_TO_VALUE, physicalCard)) {
             if (sourceToIgnore == null || modifier.getSource(gameState) == null || !Filters.and(sourceToIgnore).accepts(gameState, this, modifier.getSource(gameState))) {
